// src/types/enums/TranslationKeys.ts
export enum TranslationKeys {
  LogoSlogan = 'Logo slogan',
  Home = 'Home',
  Calendar = 'Calendar',
  Profile = 'Profile',
  Statistics = 'Statistics',
  AboutUs = 'About us',
  Exit = 'Exit',
  Start = 'Start',
  Profession = 'Profession',
  GreetingRegistration = 'Greeting registration',
  RegistrationSlogan = 'Registration slogan',
  Login = 'Login',
  CreateAccount = 'Create account',
  FirstName = 'First name',
  LastName = 'Last name',
  Email = 'Email',
  PasswordPlaceholder = 'Password placeholder',
  Registration = 'Registration',
  UseGoogle = 'Use google',
  TermsOfUseOfTheSite = 'Terms of use of the site',
  PrivacyPolicy = 'Privacy policy',
  IHaveReadThe = 'I have read the',
  AndAgreeToThem = 'And agree to them',
  AndThe = 'And the',
  SignInTo = 'Sign in to',
  ForgotYourPassword = 'Forgot your password',
  GreetingLogin = 'Greeting login',
  LoginSlogan = 'Login slogan',
}